Align all tests
