+++ lib/api/controllers/collectionController.js
@@ -230,7 +230,7 @@ function CollectionController (kuzzle) {
    * @param {Request} request
    * @returns {Promise<Object>}
    */
-  this.collectionExists = function collectionCollectionExists (request) {
+  this.exists = function collectionCollectionExists (request) {
     assertHasIndexAndCollection(request, 'collection:exist');
 
     return engine.collectionExists(request);
