+++ lib/api/controllers/realtimeController.js
@@ -3,8 +3,7 @@
 var
   assertHasBody = require('./util/requestAssertions').assertHasBody,
   assertBodyHasAttribute = require('./util/requestAssertions').assertBodyHasAttribute,
-  assertHasIndexAndCollection = require('./util/requestAssertions').assertHasIndexAndCollection,
-  assertHasId = require('./util/requestAssertions').assertHasId;
+  assertHasIndexAndCollection = require('./util/requestAssertions').assertHasIndexAndCollection;
 
 /**
  * @param {Kuzzle} kuzzle
@@ -85,18 +84,7 @@ function RealtimeController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.validate = function realtimeValidate (request) {
-    assertHasBody(request, 'realtime:validate');
-    assertHasIndexAndCollection(request, 'realtime:validate');
-    assertHasId(request, 'realtime:validate');
-
-    return kuzzle.validation.validationPromise(request, true)
-      .then(response => {
-        if (!response.valid) {
-          kuzzle.pluginsManager.trigger('validation:error', `The document does not comply with the ${request.input.resource.index} / ${request.input.resource.collection} : ${JSON.stringify(request.input.body)}`);
-        }
-
-        return Promise.resolve(response);
-      });
+    return kuzzle.funnel.controllers.document.validate(request);
   };
 }
