+++ lib/api/controllers/securityController.js
@@ -485,7 +485,7 @@ function SecurityController (kuzzle) {
         if (reset) {
           return resetRoles.call(kuzzle)
             .then(() => resetProfiles.call(kuzzle))
-            .then(() => kuzzle.funnel.controllers.index.refreshInternalIndex())
+            .then(() => kuzzle.funnel.controllers.index.refreshInternal())
             .then(() => Promise.resolve(response));
         }
