+++ lib/api/core/hotelClerk.js
@@ -94,7 +94,7 @@ function HotelClerk (kuzzle) {
           c = this.rooms[roomId].channels[channel],
           stateMatch = c.state === 'all' || !notification.state || notification.action === 'publish' || c.state === notification.state,
           scopeMatch = c.scope === 'all' || !notification.scope || c.scope === notification.scope,
-          usersMatch = c.users === 'all' || notification.controller !== 'subscribe' ||
+          usersMatch = c.users === 'all' || notification.controller !== 'realtime' ||
             c.users === 'in' && notification.action === 'on' || c.users === 'out' && notification.action === 'off';
 
         if (stateMatch && scopeMatch && usersMatch) {
