+++ lib/api/core/notifier.js
@@ -74,7 +74,7 @@ function NotifierController (kuzzle) {
     rooms = kuzzle.dsl.test(request.input.resource.index, request.input.resource.collection, request.input.body || {}, request.input.resource._id);
 
     if (rooms.length > 0) {
-      if (request.input.controller === 'write') {
+      if (request.input.controller === 'document' || request.input.controller === 'realtime') {
         notificationContent.state = 'pending';
 
         if (request.input.action === 'publish') {
