+++ lib/config/httpRoutes.js
@@ -9,14 +9,14 @@ module.exports = [
   {verb: 'get', url: '/:index/:collection/_mapping', controller: 'collection', action: 'getMapping'},
   {verb: 'get', url: '/:index/:collection/_specifications', controller: 'collection', action: 'getSpecifications'},
   {verb: 'get', url: '/users/_mapping', controller: 'collection', action: 'getUserMapping'},
-  {verb: 'get', url: '/:index/_listCollections', controller: 'collection', action: 'list'},
-  {verb: 'get', url: '/:index/_listCollections/:type', controller: 'collection', action: 'list'},
+  {verb: 'get', url: '/:index/_list', controller: 'collection', action: 'list'},
+  {verb: 'get', url: '/:index/_list/:type', controller: 'collection', action: 'list'},
 
   {verb: 'get', url: '/:index/:collection/:_id', controller: 'document', action: 'get'},
 
   {verb: 'get', url: '/:index/_autoRefresh', controller: 'index', action: 'getAutoRefresh'},
   {verb: 'get', url: '/:index/_exists', controller: 'index', action: 'exists'},
-  {verb: 'get', url: '/_listIndexes', controller: 'index', action: 'list'},
+  {verb: 'get', url: '/_list', controller: 'index', action: 'list'},
   {verb: 'get', url: '/_refreshInternal', controller: 'index', action: 'refreshInternal'},
 
   {verb: 'get', url: '/_listSubscriptions', controller: 'realtime', action: 'list'},
@@ -110,9 +110,12 @@ module.exports = [
   {verb: 'post', url: '/:index/_refresh', controller: 'index', action: 'refresh'},
   {verb: 'post', url: '/:index/_autoRefresh', controller: 'index', action: 'setAutoRefresh'},
 
-  {verb: 'post', url: '/_scroll/:scrollId', controller: 'read', action: 'scroll'},
-  {verb: 'post', url: '/:index/:collection/_search', controller: 'read', action: 'search'},
-  {verb: 'post', url: '/:index/:collection/_count', controller: 'read', action: 'count'},
+  {verb: 'post', url: '/:index/:collection/_count', controller: 'document', action: 'count'},
+  {verb: 'post', url: '/:index/:collection/_create', controller: 'document', action: 'create'},
+  {verb: 'post', url: '/:index/:collection', controller: 'document', action: 'publish'},
+  {verb: 'post', url: '/:index/:collection/_search', controller: 'document', action: 'search'},
+  {verb: 'post', url: '/_scroll/:scrollId', controller: 'document', action: 'scroll'},
+  {verb: 'post', url: '/:index/:collection/_validate', controller: 'document', action: 'validate'},
 
   {verb: 'post', url: '/_createFirstAdmin', controller: 'security', action: 'createFirstAdmin'},
   {verb: 'post', url: '/profiles/_create', controller: 'security', action: 'createProfile'},
@@ -128,9 +131,6 @@ module.exports = [
   {verb: 'post', url: '/roles/:_id', controller: 'security', action: 'updateRole'},
   {verb: 'post', url: '/users/:_id', controller: 'security', action: 'updateUser'},
 
-  {verb: 'post', url: '/:index/:collection/_create', controller: 'write', action: 'create'},
-  {verb: 'post', url: '/:index/:collection/_validate', controller: 'write', action: 'validate'},
-  {verb: 'post', url: '/:index/:collection', controller: 'write', action: 'publish'},
 
   {verb: 'post', url: '/ms/_append/:_id', controller: 'ms', action: 'append'},
   {verb: 'post', url: '/ms/_bgrewriteaof', controller: 'ms', action: 'bgrewriteaof'},
@@ -213,7 +213,7 @@ module.exports = [
   {verb: 'delete', url: '/:index/:collection/_query', controller: 'document', action: 'deleteByQuery'},
   {verb: 'delete', url: '/:index/:collection/:_id', controller: 'document', action: 'delete'},
 
-  {verb: 'delete', url: '/_deleteIndexes', controller: 'index', action: 'mDelete'},
+  {verb: 'delete', url: '/_mdelete', controller: 'index', action: 'mDelete'},
   {verb: 'delete', url: '/:index', controller: 'index', action: 'delete'},
 
   {verb: 'delete', url: '/roles/:_id', controller: 'security', action: 'deleteRole'},
