+++ test/api/core/hotelClerk/removeCustomerFromAllRooms.test.js
@@ -93,7 +93,7 @@ describe('Test: hotelClerk.removeCustomerFromAllRooms', () => {
 
           // testing requestObject argument
           should(kuzzle.notifier.notify.args[0][1]).be.instanceOf(Request);
-          should(kuzzle.notifier.notify.args[0][1].input.controller).be.exactly('subscribe');
+          should(kuzzle.notifier.notify.args[0][1].input.controller).be.exactly('realtime');
           should(kuzzle.notifier.notify.args[0][1].input.action).be.exactly('off');
           should(kuzzle.notifier.notify.args[0][1].input.resource.index).be.exactly(index);
