+++ test/api/core/hotelClerk/removeSubscription.test.js
@@ -28,7 +28,7 @@ describe('Test: hotelClerk.removeSubscription', () => {
     kuzzle.dsl = new Dsl();
 
     unsubscribeRequest = new Request({
-      controller: 'subscribe',
+      controller: 'realtime',
       action: 'off',
       index: index,
       collection: collection,
@@ -120,7 +120,7 @@ describe('Test: hotelClerk.removeSubscription', () => {
 
         // testing requestObject argument
         should(kuzzle.notifier.notify.args[0][1]).be.instanceOf(Request);
-        should(kuzzle.notifier.notify.args[0][1].input.controller).be.exactly('subscribe');
+        should(kuzzle.notifier.notify.args[0][1].input.controller).be.exactly('realtime');
         should(kuzzle.notifier.notify.args[0][1].input.action).be.exactly('off');
         should(kuzzle.notifier.notify.args[0][1].input.resource.index).be.exactly(index);
